Split up tests for clarity
